Agents/mood_json_agent/mjson_agent/requirements.txt
google-adk
google-generativeai
google-cloud-firestore
feedparser
praw
google-cloud-aiplatform[adk,agent_engines]
cloudpickle
google-cloud-trace
pydantic
